requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip#sha256=1e2d1f5da989dc1c3787c006cfb7d46a5b78bed56dbc06bfab833fd62cfcf793
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
ccxt==4.4.62
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.3
pandas==2.2.3
propcache==0.3.0
pybit==5.9.0
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
setuptools==75.8.1
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
